refactor: change types
