+++ src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const API_URL = 'https://swapi.dev/api/people';
